add `/version` command

+++ const.go
@@ -7,7 +7,7 @@ var botToken string // 全局 bot token
 var webhookURL string // Webhook 运行模式下接受请求的 URL 地址
 var webhookPort string = "localhost:2847" // Webhook 运行模式下监听的端口
 
-var logChat_ID int = -1002499888124 // 用于接收日志的聊天 ID，可以是 用户 群聊 频道
+var logChat_ID int64 = -1002499888124 // 用于接收日志的聊天 ID，可以是 用户 群聊 频道
 var logMan_IDs []int64 = []int64{1086395364, 2074319561} // 拥有查看日志权限的用户，可设定多个
 
 var metadatafile_name string = "metadata.yaml"
